# Conflicts:
#	src/main/java/com/kuuhaku/handlers/games/tabletop/games/shoukan/Hero.java
#	src/main/java/com/kuuhaku/utils/Helper.java

+++ src/main/java/com/kuuhaku/handlers/games/tabletop/games/shoukan/SlotColumn.java
@@ -21,6 +21,8 @@ package com.kuuhaku.handlers.games.tabletop.games.shoukan;
 import com.kuuhaku.handlers.games.tabletop.games.shoukan.enums.EffectTrigger;
 import com.kuuhaku.handlers.games.tabletop.games.shoukan.records.CardLink;
 
+import java.util.List;
+
 public class SlotColumn implements Cloneable {
 	private final int index;
 	private final Hand hand;
@@ -48,9 +50,10 @@ public class SlotColumn implements Cloneable {
 	public void setTop(Champion top) {
 		Champion curr = this.top;
 		if (curr != null) {
-			curr.getEffect(new EffectParameters(EffectTrigger.FINALIZE, curr.getGame(), curr.getSide(), index, Duelists.of(), curr.getGame().getChannel()));
+			if (curr.hasEffect())
+				curr.getEffect(new EffectParameters(EffectTrigger.FINALIZE, curr.getGame(), curr.getSide(), index, Duelists.of(), curr.getGame().getChannel()));
 
-			for (CardLink link : curr.getLinkedTo()) {
+			for (CardLink link : List.copyOf(curr.getLinkedTo())) {
 				curr.unlink(link.asEquipment());
 			}
 		}
@@ -70,7 +73,11 @@ public class SlotColumn implements Cloneable {
 	public void setBottom(Equipment bottom) {
 		Equipment curr = this.bottom;
 		if (curr != null) {
-			curr.getLinkedTo().asChampion().unlink(curr);
+			if (curr.hasEffect())
+				curr.getEffect(new EffectParameters(EffectTrigger.FINALIZE, curr.getGame(), curr.getSide(), index, Duelists.of(), curr.getGame().getChannel()));
+
+			if (curr.getLinkedTo() != null)
+				curr.getLinkedTo().asChampion().unlink(curr);
 		}
 
 		if (bottom != null) {
